requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.2
blinker=1.9.0=pyhff2d567_0
brotli-python=1.1.0=py313h3579c5c_2
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.1.31=hf0a4a13_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py313hc845a76_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
cryptography=44.0.2=py313h54e0d97_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
libcxx=19.1.7=ha82da77_0
libexpat=2.6.4=h286801f_0
libffi=3.4.2=h3422bc3_5
liblzma=5.6.4=h39f12f2_0
libmpdec=4.0.0=h99b78c6_0
libsqlite=3.49.1=h3f77e49_2
libzlib=1.3.1=h8359307_2
ncurses=6.5=h5e97a16_3
oauthlib=3.2.2=pyhd8ed1ab_1
openssl=3.4.1=h81ee809_0
pip=25.0.1=pyh145f28c_0
pycparser=2.22=pyh29332c3_1
pyjwt=2.10.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha55dd90_7
python=3.13.2=h81fe080_101_cp313
python-dotenv=1.0.1=pyhd8ed1ab_1
python_abi=3.13=5_cp313
readline=8.2=h1d1bf99_2
requests=2.32.3=pyhd8ed1ab_1
requests-oauthlib=2.0.0=pyhd8ed1ab_1
tk=8.6.13=h5083fa2_1
tweepy=4.15.0=pyhd8ed1ab_0
tzdata=2025a=h78e105d_0
urllib3=2.3.0=pyhd8ed1ab_0
zstandard=0.23.0=py313h90d716c_1
